requirements.txt
flask
flask-cors
mysql-connector-python
python-jose
requests
python-dotenv
